chore: Update React and React DOM versions to experimental build
This commit updates the React and React DOM versions in the package.json file to the experimental build "0.0.0-experimental-9c6806964f-20240703". This update is necessary to ensure compatibility with other dependencies and to take advantage of the latest features and improvements in React.

The commit also adds a new file, "form-submit.js", under the "components" directory. This file contains the implementation of a form submission component, which will be used in the "NewPostPage" component.

Lastly, the "NewPostPage" component is modified to import the "FormSubmit" component and the "redirect" function from the appropriate modules. The "createPost" function is updated to log the form data and store the post using the "storePost" function. After storing the post, the page is redirected to the "/feed" route.

This commit is a chore because it primarily involves updating dependencies and adding a new component, with no significant functional changes.