feat: Add form submission component and update create post functionality
The code changes include adding a new file, "form-submit.js", which contains the implementation of a form submission component. This component will be used in the "NewPostPage" component. Additionally, the "createPost" function in the "actions.js" file is updated to log the form data and store the post using the "storePost" function. After storing the post, the page is redirected to the "/feed" route.

This commit introduces a new feature and improves the functionality of creating posts.